Added Reputation Check
The Reputation is now checked upon question creation. The amount of
reputation needed and gained by specific actions is now stored in the
database. Upon setup, call manage.py load_reputation to load the
default key/value pairs for the actions into the db. from then on, you
can change the values with the admin panel