Added Closed feature for Questions
Kind of a big commit, so here it comes:

1. Questions can now be closed, from which point onward they can not be
voted for anymore. The question with the most votes each day should be
closed for answering.
2. You can filter for answered/unanswered in /Questions/ and
/Questions/my/ by appending ?answered=true|false to the url. This will
filter them based on the "closed" property
3. There is a new Management command, which gets the question with the
most upvotes, and sends it to slack. i'll setup a cron job so it does
this every day at 12 pm.
4. The Admin panel now can also sort questions based on upvotes. Also,
you can select multiple questions and choose "close questions" from the
dropdown menu to close them all.